@@ -1,16 +1,16 @@
 {
-    "name": "detailnet/denner-client-module",
-    "description": "Zend Framework Module for Denner-client",
+    "name": "detailnet/jumbo-client-module",
+    "description": "Zend Framework Module for jumbo-client",
     "type": "library",
     "keywords": [
-        "denner",
+        "jumbo",
         "detailnet",
         "api",
         "client",
         "module",
         "zf"
     ],
-    "homepage": "https://github.com/detailnet/denner-client-module/",
+    "homepage": "https://github.com/detailnet/jumbo-client-module/",
     "license": "MIT",
     "authors": [
         {
@@ -31,15 +31,15 @@
         }
     ],
     "require": {
-        "php": "^5.5 || ^7.0",
+        "php": "^7.0",
         "zendframework/zend-loader": "^2.5.1",
         "zendframework/zend-modulemanager": "^2.7.2",
         "zendframework/zend-servicemanager": "^2.7.8",
         "detailnet/dfw-core": "^0.3",
-        "detailnet/denner-client": "^1.0"
+        "detailnet/jumbo-client": "^0.1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.6",
+        "phpunit/phpunit": "^6.1",
         "squizlabs/php_codesniffer" : "^2.2",
         "phpmd/phpmd": "^2.2",
         "satooshi/php-coveralls": "^1.0"
@@ -48,12 +48,12 @@
     },
     "autoload": {
         "psr-0": {
-            "Denner\\Client\\": "src/"
+            "Jumbo\\Client\\": "src/"
         }
     },
     "autoload-dev": {
         "psr-0": {
-            "Denner\\Client\\": "tests/"
+            "Jumbo\\Client\\": "tests/"
         }
     },
     "extra": {
