@@ -4,4 +4,4 @@
  * It allows usage of this module even without composer.
  * The original Module.php is in 'src/Detail/Locale' in order to respect PSR-0.
  */
-require_once __DIR__ . '/src/Denner/Client/Module.php';
+require_once __DIR__ . '/src/Jumbo/Client/Module.php';
