@@ -3,21 +3,21 @@
 return array(
     'service_manager' => array(
         'abstract_factories' => array(
-            'Denner\Client\Factory\Client\ClientFactory',
+            'Jumbo\Client\Factory\Client\ClientFactory',
         ),
         'aliases' => array(
         ),
         'invokables' => array(
         ),
         'factories' => array(
-            'Denner\Client\Options\ModuleOptions' => 'Denner\Client\Factory\Options\ModuleOptionsFactory',
+            'Jumbo\Client\Options\ModuleOptions' => 'Jumbo\Client\Factory\Options\ModuleOptionsFactory',
         ),
         'initializers' => array(
         ),
         'shared' => array(
         ),
     ),
-    'denner_client' => array(
+    'jumbo_client' => array(
         'clients' => array(
         ),
     ),
