@@ -9,7 +9,7 @@ if (file_exists($basePath . 'vendor/autoload.php')) {
     throw new RuntimeException('vendor/autoload.php could not be found. Did you run `php composer.phar install`?');
 }
 
-$loader->add('DennerTest\Client', __DIR__);
+$loader->add('JumboTest\Client', __DIR__);
 
 if (!$config = @include $basePath . 'tests/configuration.php') {
     $config = require $basePath . 'tests/configuration.php.dist';
